Thunder tech fully implemented + bugfixes

+++ Assets/ScriptableObjects/BuffEffects/BuffEffectSO.cs
@@ -8,5 +8,5 @@ public class BuffEffectSO : ScriptableObject
     [field: SerializeField] public int MaxStacks { get; private set; }
     [field: SerializeField] public float Duration { get; private set; }
     [field: SerializeField] public Sprite Icon { get; private set; }
-    [field: SerializeField] public List<SerializedPair<PlayerStatGroup.Stat, int>> Effects { get; private set; }
+    [field: SerializeField] public List<SerializedPair<PlayerStatGroup.Stat, float>> Effects { get; private set; }
 }
